requirements.txt
adbc-driver-manager==1.1.0
adbc-driver-sqlite==0.8.0
aiobotocore==2.13.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.5.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==24.1.0
babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.131
botocore==1.34.131
broad-babel==0.1.22
captum==0.7.0
cattrs==23.2.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
copairs==0.4.1
cupy-cuda12x==13.2.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
docutils==0.21.2
executing==2.0.1
fasteners==0.19
fastjsonschema==2.20.0
fastrlock==0.8.2
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperopt==0.2.7
idna==3.7
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jump-portrait==0.0.22
jupyter==1.0.0
jupyter-client==8.6.2
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.11
kiwisolver==1.4.5
lightning==2.4.0
lightning-utilities==0.11.6
llvmlite==0.43.0
lsprotocol==2023.0.1
marimo==0.9.27
markdown==3.7
markupsafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
munch==4.0.0
narwhals==1.14.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook-shim==0.2.4
numba==0.60.0
numcodecs==0.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
plotly==5.23.0
polars==1.16.0
pooch==1.7.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.7
pyarrow==16.1.0
pycparser==2.22
pyflakes==3.2.0
pygls==1.3.1
pygments==2.18.0
pymdown-extensions==10.12
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.4.0
pytz==2024.1
pyyaml==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
qtpy==2.4.1
ray==2.34.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
ruff==0.5.6
ruff-lsp==0.0.54
s3fs==2024.6.1
s3path==0.5.7
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
send2trash==1.8.3
setuptools==65.5.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.3
sympy==1.13.2
tenacity==9.0.0
tensorboardx==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomlkit==0.13.2
torch==2.4.0
torch-fidelity==0.3.0
torcheval==0.0.7
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
tune-sklearn==0.5.0
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.11
wrapt==1.16.0
xarray==2024.9.0
xgboost==2.1.1
yarl==1.9.4
zarr==2.18.3
